Fix port flood rate truncation and migration silent failures
- Port flood: use uint64 multiply before divide to reduce integer
  truncation (e.g. hits=1,seconds=120 no longer rounds to 0)
- Migration --apply: log first error and count failures for blocked/allowed
  IP restoration instead of silently ignoring all errors. Reports totals
  and warns about deny_ip_limit if failures occur.

+++ internal/firewall/engine.go
@@ -420,8 +420,8 @@ func (e *Engine) createInputChain() error {
 		if pf.Proto == "udp" {
 			proto = 17
 		}
-		// Convert hits/seconds to per-minute rate
-		ratePerMin := uint64(pf.Hits * 60 / pf.Seconds)
+		// Convert hits/seconds to per-minute rate (multiply first to reduce truncation)
+		ratePerMin := uint64(pf.Hits) * 60 / uint64(pf.Seconds)
 		if ratePerMin < 1 {
 			ratePerMin = 1
 		}
